refactor: split engine state into focused files
