+++ chessTest/internal/game/engine.go
@@ -1,3 +1,4 @@
+// Package game implements the core battle chess engine state and API.
 package game
 
 import (
@@ -6,10 +7,6 @@ import (
 
 	"battle_chess_poc/internal/shared"
 )
-
-// ---------------------------
-// Core Engine & State Structs
-// ---------------------------
 
 // Engine encapsulates the optimized chess engine with ability metadata.
 type Engine struct {
@@ -24,86 +21,6 @@ type Engine struct {
 	pendingDoOver map[int]bool // Tracks per-piece DoOver consumption
 	currentMove   *MoveState
 	temporalSlow  map[Color]int
-}
-
-// undoState captures a snapshot of the engine's state for history.
-type undoState struct {
-	board         Board
-	blockFacing   map[int]Direction
-	lastNote      string
-	locked        bool
-	configured    map[Color]bool
-	pendingDoOver map[int]bool
-	currentMove   *MoveState
-	temporalSlow  map[Color]int
-}
-
-func cloneMoveState(src *MoveState, pieceMap map[*Piece]*Piece) *MoveState {
-	if src == nil {
-		return nil
-	}
-	clone := *src
-	clone.Path = append([]Square(nil), src.Path...)
-	clone.Captures = make([]*Piece, len(src.Captures))
-	for i, captured := range src.Captures {
-		clone.Captures[i] = mapPiecePointer(captured, pieceMap)
-	}
-	clone.Piece = mapPiecePointer(src.Piece, pieceMap)
-	return &clone
-}
-
-func mapPiecePointer(pc *Piece, pieceMap map[*Piece]*Piece) *Piece {
-	if pc == nil {
-		return nil
-	}
-	if mapped, ok := pieceMap[pc]; ok {
-		return mapped
-	}
-	return clonePiece(pc)
-}
-
-func cloneIntDirectionMap(src map[int]Direction) map[int]Direction {
-	if len(src) == 0 {
-		return make(map[int]Direction)
-	}
-	clone := make(map[int]Direction, len(src))
-	for k, v := range src {
-		clone[k] = v
-	}
-	return clone
-}
-
-func cloneColorBoolMap(src map[Color]bool) map[Color]bool {
-	if len(src) == 0 {
-		return make(map[Color]bool)
-	}
-	clone := make(map[Color]bool, len(src))
-	for k, v := range src {
-		clone[k] = v
-	}
-	return clone
-}
-
-func cloneIntBoolMap(src map[int]bool) map[int]bool {
-	if len(src) == 0 {
-		return make(map[int]bool)
-	}
-	clone := make(map[int]bool, len(src))
-	for k, v := range src {
-		clone[k] = v
-	}
-	return clone
-}
-
-func cloneColorIntMap(src map[Color]int) map[Color]int {
-	if len(src) == 0 {
-		return make(map[Color]int)
-	}
-	clone := make(map[Color]int, len(src))
-	for k, v := range src {
-		clone[k] = v
-	}
-	return clone
 }
 
 // Board represents the state of the chessboard.
@@ -179,10 +96,6 @@ type BoardState struct {
 	EnPassant        EnPassantTarget     `json:"enPassant"`
 	PromotionChoices PromotionChoices    `json:"promotionChoices"`
 }
-
-// ---------------------------
-// Public API
-// ---------------------------
 
 // NewEngine creates and initializes a new game engine.
 func NewEngine() *Engine {
@@ -370,511 +283,6 @@ func (e *Engine) State() BoardState {
 	return state
 }
 
-// ---------------------------
-// Move Lifecycle Management
-// ---------------------------
-
-// ---------------------------
-// Ability Resolution Pipeline
-// ---------------------------
-
-// ResolveCaptureAbility handles special ability triggers on capture.
-func (e *Engine) ResolveCaptureAbility(attacker, victim *Piece, captureSquare Square) error {
-	if victim == nil {
-		return nil
-	}
-	if err := e.maybeTriggerDoOver(victim); err != nil {
-		return err
-	}
-
-	victimRank := rankOf(victim.Type)
-	victimColor := victim.Color
-
-	extraRemoved := false
-	if attacker != nil && attacker.Abilities.Contains(AbilityDoubleKill) {
-		if target := e.trySmartExtraCapture(attacker, captureSquare, victimColor, victimRank); target != nil {
-			targetSquare := target.Square
-			if removed, err := e.attemptAbilityRemoval(attacker, target); err != nil {
-				return err
-			} else if removed {
-				appendAbilityNote(&e.board.lastNote, fmt.Sprintf("DoubleKill: removed %s %s at %s", target.Color, target.Type, targetSquare))
-				extraRemoved = true
-			}
-		}
-	}
-
-	if !extraRemoved && attacker != nil && elementOf(e, attacker) == ElementFire && attacker.Abilities.Contains(AbilityScorch) {
-		if target := e.trySmartExtraCapture(attacker, captureSquare, victimColor, victimRank); target != nil {
-			targetSquare := target.Square
-			if removed, err := e.attemptAbilityRemoval(attacker, target); err != nil {
-				return err
-			} else if removed {
-				appendAbilityNote(&e.board.lastNote, fmt.Sprintf("Fire Scorch: removed %s %s at %s", target.Color, target.Type, targetSquare))
-			}
-		}
-	}
-
-	if attacker != nil && e.currentMove != nil && e.currentMove.HasQuantumKill && !e.currentMove.QuantumKillUsed {
-		e.currentMove.QuantumKillUsed = true
-		if target := e.findQuantumKillTarget(attacker, victimColor, victimRank); target != nil {
-			targetSquare := target.Square
-			if removed, err := e.attemptAbilityRemoval(attacker, target); err != nil {
-				return err
-			} else if removed {
-				appendAbilityNote(&e.board.lastNote, fmt.Sprintf("Quantum Kill: removed %s %s at %s", target.Color, target.Type, targetSquare))
-				if echo := e.trySmartExtraCapture(attacker, targetSquare, victimColor, rankOf(target.Type)); echo != nil {
-					echoSquare := echo.Square
-					if removedEcho, err := e.attemptAbilityRemoval(attacker, echo); err != nil {
-						return err
-					} else if removedEcho {
-						appendAbilityNote(&e.board.lastNote, fmt.Sprintf("Quantum Echo: removed %s %s at %s", echo.Color, echo.Type, echoSquare))
-					}
-				}
-			}
-		}
-	}
-
-	e.applyCapturePenalties(attacker)
-	return nil
-}
-
-// trySmartExtraCapture scans neighbors and identifies the highest-rank, eligible, lower-rank piece.
-func (e *Engine) trySmartExtraCapture(attacker *Piece, captureSquare Square, victimColor Color, victimRank int) *Piece {
-	var bestP *Piece
-	bestRank := -1
-
-	for _, d := range [...]int{-9, -8, -7, -1, 1, 7, 8, 9} {
-		nsqInt := int(captureSquare) + d
-		if nsqInt < 0 || nsqInt >= 64 {
-			continue
-		}
-		nsq := Square(nsqInt)
-		if !shared.SameFileNeighborhood(captureSquare, nsq, d) {
-			continue
-		}
-		p := e.board.pieceAt[nsq]
-		if p == nil || p.Color != victimColor {
-			continue
-		}
-		if !e.canAbilityRemove(attacker, p) {
-			continue
-		}
-		if elementOf(e, p) == ElementEarth || p.Abilities.Contains(AbilityObstinant) || e.abilities[p.Color].Contains(AbilityObstinant) {
-			continue
-		}
-		r := rankOf(p.Type)
-		if r < victimRank && r > bestRank {
-			bestRank = r
-			bestP = p
-		}
-	}
-
-	return bestP
-}
-
-func (e *Engine) maybeTriggerDoOver(victim *Piece) error {
-	if victim == nil || !victim.Abilities.Contains(AbilityDoOver) || e.pendingDoOver[victim.ID] {
-		return nil
-	}
-
-	plies := 4
-	if plies > len(e.history) {
-		plies = len(e.history)
-	}
-	if plies > 0 {
-		e.popHistory(plies)
-		victim.Abilities = victim.Abilities.Without(AbilityDoOver)
-		e.pendingDoOver[victim.ID] = true
-		e.board.lastNote = fmt.Sprintf("DoOver: %s %s rewound %d plies (%.1f turns)", victim.Color, victim.Type, plies, float64(plies)/2.0)
-		return ErrDoOverActivated
-	}
-
-	victim.Abilities = victim.Abilities.Without(AbilityDoOver)
-	e.pendingDoOver[victim.ID] = true
-	return nil
-}
-
-func (e *Engine) canAbilityRemove(attacker, target *Piece) bool {
-	if target == nil {
-		return false
-	}
-	if target.Type == King {
-		return false
-	}
-	if target.Abilities.Contains(AbilityIndomitable) {
-		return false
-	}
-	if target.Abilities.Contains(AbilityStalwart) && attacker != nil && rankOf(attacker.Type) < rankOf(target.Type) {
-		return false
-	}
-	if target.Abilities.Contains(AbilityBelligerent) && attacker != nil && rankOf(attacker.Type) > rankOf(target.Type) {
-		return false
-	}
-	return true
-}
-
-func (e *Engine) attemptAbilityRemoval(attacker, target *Piece) (bool, error) {
-	if target == nil || !e.canAbilityRemove(attacker, target) {
-		return false, nil
-	}
-
-	e.removePiece(target, target.Square)
-	if err := e.maybeTriggerDoOver(target); err != nil {
-		return true, err
-	}
-	return true, nil
-}
-
-func (e *Engine) findQuantumKillTarget(attacker *Piece, victimColor Color, maxRank int) *Piece {
-	var best *Piece
-	bestRank := -1
-	bestIndex := 65
-
-	for idx, p := range e.board.pieceAt {
-		if p == nil || p.Color != victimColor {
-			continue
-		}
-		rank := rankOf(p.Type)
-		if rank > maxRank {
-			continue
-		}
-		if !e.canAbilityRemove(attacker, p) {
-			continue
-		}
-		if rank > bestRank || (rank == bestRank && idx < bestIndex) {
-			best = p
-			bestRank = rank
-			bestIndex = idx
-		}
-	}
-
-	return best
-}
-
-func (e *Engine) applyCapturePenalties(attacker *Piece) {
-	if e.currentMove == nil || attacker == nil {
-		return
-	}
-
-	element := elementOf(e, attacker)
-	if attacker.Abilities.Contains(AbilityPoisonousMeat) && element != ElementShadow {
-		if e.currentMove.RemainingSteps > 0 {
-			e.currentMove.RemainingSteps--
-			appendAbilityNote(&e.board.lastNote, "Poisonous Meat drains 1 step")
-		}
-	}
-
-	if attacker.Abilities.Contains(AbilityOverload) && element == ElementLightning && attacker.Abilities.Contains(AbilityStalwart) {
-		if e.currentMove.RemainingSteps > 0 {
-			e.currentMove.RemainingSteps--
-			appendAbilityNote(&e.board.lastNote, "Overload + Stalwart costs 1 step")
-		}
-	}
-}
-
-// ---------------------------
-// Step, Cost & State Calculation
-// ---------------------------
-
-// ---------------------------
-// History & State Management
-// ---------------------------
-
-func (e *Engine) snapshot() undoState {
-	boardClone, mapping := e.board.cloneWithMap()
-	var moveCopy *MoveState
-	if e.currentMove != nil {
-		moveCopy = cloneMoveState(e.currentMove, mapping)
-	}
-	s := undoState{
-		board:         boardClone,
-		blockFacing:   cloneIntDirectionMap(e.blockFacing),
-		lastNote:      e.board.lastNote,
-		locked:        e.locked,
-		configured:    cloneColorBoolMap(e.configured),
-		pendingDoOver: cloneIntBoolMap(e.pendingDoOver),
-		currentMove:   moveCopy,
-		temporalSlow:  cloneColorIntMap(e.temporalSlow),
-	}
-	return s
-}
-
-func (e *Engine) applySnapshot(s undoState) {
-	boardClone, mapping := s.board.cloneWithMap()
-	e.board = boardClone
-	e.blockFacing = cloneIntDirectionMap(s.blockFacing)
-	e.board.lastNote = s.lastNote
-	e.locked = s.locked
-	e.configured = cloneColorBoolMap(s.configured)
-	e.pendingDoOver = cloneIntBoolMap(s.pendingDoOver)
-	e.temporalSlow = cloneColorIntMap(s.temporalSlow)
-	if s.currentMove != nil {
-		e.currentMove = cloneMoveState(s.currentMove, mapping)
-	} else {
-		e.currentMove = nil
-	}
-}
-
-func (e *Engine) pushHistory() { e.history = append(e.history, e.snapshot()) }
-func (e *Engine) popHistory(n int) {
-	for i := 0; i < n && len(e.history) > 0; i++ {
-		idx := len(e.history) - 1
-		s := e.history[idx]
-		e.history = e.history[:idx]
-		e.applySnapshot(s)
-	}
-}
-
-// ---------------------------
-// Legality & Validation Helpers
-// ---------------------------
-
-func (e *Engine) isPathClear(from, to Square) bool {
-	line := shared.Line(from, to)
-	for _, sq := range line {
-		if e.board.pieceAt[sq] != nil {
-			return false
-		}
-	}
-	return true
-}
-
-func (e *Engine) canPhaseThrough(pc *Piece, _ Square, _ Square) bool {
-	if pc == nil {
-		return false
-	}
-
-	hasFlood := pc.Abilities.Contains(AbilityFloodWake)
-	hasBastion := pc.Abilities.Contains(AbilityBastion)
-	hasGale := pc.Abilities.Contains(AbilityGaleLift)
-
-	if e.abilities != nil {
-		if al, ok := e.abilities[pc.Color]; ok {
-			if al.Contains(AbilityFloodWake) {
-				hasFlood = true
-			}
-			if al.Contains(AbilityBastion) {
-				hasBastion = true
-			}
-			if al.Contains(AbilityGaleLift) {
-				hasGale = true
-			}
-		}
-	}
-
-	if hasFlood || hasBastion {
-		return false
-	}
-	if hasGale {
-		return true
-	}
-	if pc.Abilities.Contains(AbilityUmbralStep) {
-		return true
-	}
-	if e.abilities != nil {
-		if al, ok := e.abilities[pc.Color]; ok && al.Contains(AbilityUmbralStep) {
-			return true
-		}
-	}
-	return false
-}
-
-// ---------------------------
-// General Helpers
-// ---------------------------
-
-func appendAbilityNote(dst *string, note string) {
-	if *dst == "" || *dst == "New game" || *dst == "Configuration locked - game started" {
-		*dst = note
-	} else {
-		*dst += "; " + note
-	}
-}
-
-type moveSegmentContext struct {
-	capture       *Piece
-	captureSquare Square
-	enPassant     bool
-}
-
-func (e *Engine) executeMoveSegment(from, to Square, ctx moveSegmentContext) {
-	pc := e.board.pieceAt[from]
-	if pc == nil {
-		return
-	}
-
-	isCastle := pc.Type == King && from.Rank() == to.Rank() && absInt(to.File()-from.File()) == 2
-
-	if ctx.capture != nil {
-		e.updateCastlingRightsForCapture(ctx.capture, ctx.captureSquare)
-		e.removePiece(ctx.capture, ctx.captureSquare)
-		if ctx.enPassant {
-			appendAbilityNote(&e.board.lastNote, "En passant capture")
-		}
-	}
-
-	e.board.EnPassant = NoEnPassantTarget()
-
-	e.board.pieceAt[from] = nil
-	pc.Square = to
-	e.board.pieceAt[to] = pc
-
-	e.board.pieces[pc.Color][pc.Type] = e.board.pieces[pc.Color][pc.Type].Remove(from).Add(to)
-	e.board.occupancy[pc.Color] = e.board.occupancy[pc.Color].Remove(from).Add(to)
-	e.board.allOcc = e.board.allOcc.Remove(from).Add(to)
-
-	e.updateCastlingRightsForMove(pc, from)
-
-	if pc.Type == Pawn {
-		diff := to.Rank() - from.Rank()
-		if diff == 2 || diff == -2 {
-			midRank := from.Rank() + diff/2
-			if sq, ok := shared.SquareFromCoords(midRank, from.File()); ok {
-				e.board.EnPassant = NewEnPassantTarget(sq)
-			}
-		}
-	}
-
-	if isCastle {
-		e.performCastleRookMove(pc.Color, from, to)
-	}
-}
-
-func (e *Engine) placePiece(color Color, pt PieceType, sq Square) {
-	id := e.nextPieceID
-	e.nextPieceID++
-	pc := &Piece{
-		ID:        id,
-		Color:     color,
-		Type:      pt,
-		Square:    sq,
-		Abilities: e.abilities[color].Clone(),
-		Element:   e.elements[color],
-		BlockDir:  DirNone,
-	}
-	e.board.pieceAt[sq] = pc
-	e.board.pieces[color][pt] = e.board.pieces[color][pt].Add(sq)
-	e.board.occupancy[color] = e.board.occupancy[color].Add(sq)
-	e.board.allOcc = e.board.allOcc.Add(sq)
-}
-
-func castlingRightForRook(color Color, sq Square) CastlingRights {
-	switch color {
-	case White:
-		if sq.Rank() != 0 {
-			return CastlingNone
-		}
-		switch sq.File() {
-		case 0:
-			return CastlingRight(White, CastleQueenside)
-		case 7:
-			return CastlingRight(White, CastleKingside)
-		}
-	case Black:
-		if sq.Rank() != 7 {
-			return CastlingNone
-		}
-		switch sq.File() {
-		case 0:
-			return CastlingRight(Black, CastleQueenside)
-		case 7:
-			return CastlingRight(Black, CastleKingside)
-		}
-	}
-	return CastlingNone
-}
-
-func (e *Engine) updateCastlingRightsForMove(pc *Piece, from Square) {
-	if pc == nil {
-		return
-	}
-	switch pc.Type {
-	case King:
-		e.board.Castling = e.board.Castling.WithoutColor(pc.Color)
-	case Rook:
-		e.board.Castling = e.board.Castling.Without(castlingRightForRook(pc.Color, from))
-	}
-}
-
-func (e *Engine) updateCastlingRightsForCapture(pc *Piece, sq Square) {
-	if pc == nil {
-		return
-	}
-	switch pc.Type {
-	case King:
-		e.board.Castling = e.board.Castling.WithoutColor(pc.Color)
-	case Rook:
-		e.board.Castling = e.board.Castling.Without(castlingRightForRook(pc.Color, sq))
-	}
-}
-
-func (e *Engine) performCastleRookMove(color Color, from, to Square) {
-	rank := from.Rank()
-	var rookFromFile, rookToFile int
-	var note string
-	if to.File() > from.File() {
-		rookFromFile = 7
-		rookToFile = to.File() - 1
-		note = "Castled kingside"
-	} else {
-		rookFromFile = 0
-		rookToFile = to.File() + 1
-		note = "Castled queenside"
-	}
-	rookFrom, okFrom := shared.SquareFromCoords(rank, rookFromFile)
-	rookTo, okTo := shared.SquareFromCoords(rank, rookToFile)
-	if !okFrom || !okTo {
-		return
-	}
-	rook := e.board.pieceAt[rookFrom]
-	if rook == nil || rook.Type != Rook || rook.Color != color {
-		return
-	}
-
-	e.board.pieceAt[rookFrom] = nil
-	rook.Square = rookTo
-	e.board.pieceAt[rookTo] = rook
-
-	e.board.pieces[color][Rook] = e.board.pieces[color][Rook].Remove(rookFrom).Add(rookTo)
-	e.board.occupancy[color] = e.board.occupancy[color].Remove(rookFrom).Add(rookTo)
-	e.board.allOcc = e.board.allOcc.Remove(rookFrom).Add(rookTo)
-
-	appendAbilityNote(&e.board.lastNote, note)
-}
-
-func (e *Engine) removePiece(pc *Piece, sq Square) {
-	e.board.pieces[pc.Color][pc.Type] = e.board.pieces[pc.Color][pc.Type].Remove(sq)
-	e.board.occupancy[pc.Color] = e.board.occupancy[pc.Color].Remove(sq)
-	e.board.allOcc = e.board.allOcc.Remove(sq)
-	e.board.pieceAt[sq] = nil
-	delete(e.blockFacing, pc.ID)
-	delete(e.pendingDoOver, pc.ID)
-}
-
-func (b *Board) cloneWithMap() (Board, map[*Piece]*Piece) {
-	out := *b
-	mapping := make(map[*Piece]*Piece, len(b.pieceAt))
-	for i, pc := range b.pieceAt {
-		if pc != nil {
-			copyPc := clonePiece(pc)
-			out.pieceAt[i] = copyPc
-			mapping[pc] = copyPc
-		} else {
-			out.pieceAt[i] = nil
-		}
-	}
-	return out, mapping
-}
-
-func clonePiece(pc *Piece) *Piece {
-	if pc == nil {
-		return nil
-	}
-	clone := *pc
-	clone.Abilities = pc.Abilities.Clone()
-	return &clone
-}
-
 func (e *Engine) flipTurn() { e.board.turn = e.board.turn.Opposite() }
 
 func (e *Engine) ensureConfigured() error {
@@ -924,20 +332,6 @@ func (e *Engine) resolveBlockPathFacing(pc *Piece, dir Direction) string {
 	return fmt.Sprintf("BlockPath facing %s", dir)
 }
 
-func (e *Engine) captureBlockedByBlockPath(attacker *Piece, from Square, defender *Piece, to Square) (bool, string) {
-	if defender == nil || !defender.Abilities.Contains(AbilityBlockPath) || defender.BlockDir == DirNone {
-		return false, ""
-	}
-	if attacker != nil && elementOf(e, attacker) == ElementWater {
-		return false, ""
-	}
-	dir := shared.DirectionOf(from, to)
-	if dir == defender.BlockDir {
-		return true, fmt.Sprintf("Capture blocked by BlockPath (%s)", defender.BlockDir)
-	}
-	return false, ""
-}
-
 func (e *Engine) resolvePromotion(pc *Piece) {
 	if pc.Type != Pawn {
 		return
@@ -962,19 +356,18 @@ func (e *Engine) selectPromotionPiece(color Color) PieceType {
 	return choices.Default()
 }
 
-// ---------------------------
-// Data-driven & Static Helpers
-// ---------------------------
-
 func (e *Engine) hasChainKill(p *Piece) bool {
 	return p != nil && p.Abilities.Contains(AbilityChainKill)
 }
+
 func (e *Engine) hasQuantumKill(p *Piece) bool {
 	return p != nil && p.Abilities.Contains(AbilityQuantumKill)
 }
+
 func (e *Engine) hasDoubleKill(p *Piece) bool {
 	return p != nil && p.Abilities.Contains(AbilityDoubleKill)
 }
+
 func (e *Engine) isSlider(pt PieceType) bool { return pt == Queen || pt == Rook || pt == Bishop }
 
 var RankOrder = map[PieceType]int{King: 5, Queen: 4, Rook: 3, Bishop: 2, Knight: 2, Pawn: 1}
